Train time distortion maps
Horrible, nasty code I wrote in 2013 that produced the beautiful site at
http://tomforth.co.uk/traintimes/

My suggestion would be to rewrite completely but make use of the
examples in distortbatch.bat which will run on any OS shell as a guide.
The Shepards distortion is extremely powerful and the parameters took a
while to get right.

I'm a much better programmer now than when I wrote this.